refactor: extract to useCommonParamsOfBlock
